Fix merge issue

+++ precompile/staking/src/lib.rs
@@ -53,11 +53,11 @@ where
 	<<Runtime as frame_system::Config>::RuntimeCall as Dispatchable>::RuntimeOrigin: OriginTrait,
 	<<Runtime as darwinia_staking::Config>::Deposit as Stake>::Item: From<u16>,
 {
-	#[precompile::public("stake(uint256,uint8[])")]
+	#[precompile::public("stake(uint256,uint16[])")]
 	fn stake(
 		handle: &mut impl PrecompileHandle,
 		ring_amount: U256,
-		deposits: Vec<u8>,
+		deposits: Vec<u16>,
 	) -> EvmResult<bool> {
 		let origin = handle.context().caller.into();
 		let deposits = deposits.into_iter().map(|i| i.into()).collect();
@@ -73,11 +73,11 @@ where
 		Ok(true)
 	}
 
-	#[precompile::public("unstake(uint256,uint8[])")]
+	#[precompile::public("unstake(uint256,uint16[])")]
 	fn unstake(
 		handle: &mut impl PrecompileHandle,
 		ring_amount: U256,
-		deposits: Vec<u8>,
+		deposits: Vec<u16>,
 	) -> EvmResult<bool> {
 		let origin = handle.context().caller.into();
 		let deposits = deposits.into_iter().map(|i| i.into()).collect();
@@ -93,11 +93,11 @@ where
 		Ok(true)
 	}
 
-	#[precompile::public("restake(uint256,uint8[])")]
+	#[precompile::public("restake(uint256,uint16[])")]
 	fn restake(
 		handle: &mut impl PrecompileHandle,
 		ring_amount: U256,
-		deposits: Vec<u8>,
+		deposits: Vec<u16>,
 	) -> EvmResult<bool> {
 		let origin = handle.context().caller.into();
 		let deposits = deposits.into_iter().map(|i| i.into()).collect();
